fix aut test fialures
