+++ openmetadata-ui/src/main/resources/ui/cypress/common/VersionUtils.js
@@ -18,8 +18,8 @@ import {
   interceptURL,
   verifyResponseStatusCode,
   visitDataModelPage,
-  visitEntityDetailsPage,
 } from './common';
+import { visitEntityDetailsPage } from './Utils/Entity';
 
 export const visitEntityDetailsVersionPage = (
   entityDetails,
